Move configuration files to new config directory
